backend-django/pyproject.toml
[project]
name = "backend-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-shortcuts>=1.6",
    "djangorestframework>=3.16.0",
    "djangorestframework-jwt>=1.11.0",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.2",
    "djangorestframework-stubs>=3.16.1",
    "types-requests>=2.32.4.20250611",
]
